fix datasource bug
