+++ dinky-web/src/pages/RegCenter/Alert/AlertTemplate/index.tsx
@@ -21,7 +21,7 @@ import { EditBtn } from '@/components/CallBackButton/EditBtn';
 import { NormalDeleteBtn } from '@/components/CallBackButton/NormalDeleteBtn';
 import { Authorized } from '@/hooks/useAccess';
 import AlertTemplateForm from '@/pages/RegCenter/Alert/AlertTemplate/components/AlertTemplateForm';
-import { handleAddOrUpdate, handleRemoveById } from '@/services/BusinessCrud';
+import { handleAddOrUpdate,handleRemoveById } from '@/services/BusinessCrud';
 import { API_CONSTANTS } from '@/services/endpoints';
 import { Alert } from '@/types/RegCenter/data';
 import { InitAlertTemplateState } from '@/types/RegCenter/init.d';
@@ -30,17 +30,15 @@ import { l } from '@/utils/intl';
 import { useRequest } from '@@/exports';
 import { PlusOutlined } from '@ant-design/icons';
 import { PageContainer } from '@ant-design/pro-layout';
-import { Button, Card, List, Modal, Typography } from 'antd';
+import { Button,Card,List,Modal } from 'antd';
 import { useState } from 'react';
 import Markdown from 'react-markdown';
-const { Paragraph } = Typography;
 
 export default () => {
   const [alertTemplateState, setAlertTemplateState] =
     useState<AlertTemplateState>(InitAlertTemplateState);
 
   const { data, loading, run } = useRequest({ url: API_CONSTANTS.ALERT_TEMPLATE });
-  const nullData: Partial<Alert.AlertTemplate> = {};
 
   /**
    * edit click callback
@@ -151,11 +149,24 @@ export default () => {
 
   return (
     <PageContainer>
+      <Button
+        style={{ width: '100%', margin: '10px 0' }}
+        type={'dashed'}
+        icon={<PlusOutlined />}
+        onClick={() =>
+          setAlertTemplateState((prevState) => ({
+            ...prevState,
+            addedOpen: true
+          }))
+        }
+      >
+        {l('button.create')}
+      </Button>
       <List<Alert.AlertTemplate>
         rowKey='id'
         loading={loading}
         grid={{ gutter: 16, xs: 1, sm: 2, md: 3, lg: 3, xl: 4, xxl: 4 }}
-        dataSource={[nullData, ...(data ?? [])]}
+        dataSource={data ?? []}
         renderItem={(item) => renderTemplateCard(item)}
       />
